add files store by id list, check bind to other estimation before delete

+++ src/main/java/ru/irlix/evaluation/dao/helper/FileStorageHelper.java
@@ -72,6 +72,24 @@ public class FileStorageHelper {
         fileStorageRepository.saveAll(fileStorageList);
     }
 
+    public void storeFileListByIds(List<Long> idList, Estimation estimation) {
+        List<FileStorage> existedFileStorageList = fileStorageRepository.findByIdIn(idList);
+        List<FileStorage> copiedFileStorageList = new ArrayList<>();
+        existedFileStorageList.forEach(file -> {
+            if (checkExist(file)) {
+                FileStorage fileStorage = FileStorage.builder()
+                        .uuid(file.getUuid())
+                        .fileName(file.getFileName())
+                        .docType(file.getDocType())
+                        .folder(file.getFolder())
+                        .estimation(estimation)
+                        .build();
+                copiedFileStorageList.add(fileStorage);
+            }
+        });
+        fileStorageRepository.saveAll(copiedFileStorageList);
+    }
+
     @LogInfo
     public List<FileResponse> getFileResponseList(Estimation estimation) {
         List<FileResponse> fileResponseList = new ArrayList<>();
@@ -100,13 +118,19 @@ public class FileStorageHelper {
         String extension = fileName.substring(fileName.lastIndexOf("."));
         Path filePath = rootLocation.resolve(fileStorage.getUuid().toString() + extension).normalize();
         if (Files.exists(filePath))
-            try {
-                Files.delete(filePath);
-                log.info("File with id " + fileStorage.getId() + " deleted");
-            } catch (IOException e) {
-                e.printStackTrace();
-                log.error(e.getMessage());
+            if (!checkBind(fileStorage)) {
+                try {
+                    Files.delete(filePath);
+                    log.info("File with id " + fileStorage.getId() + " deleted");
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    log.error(e.getMessage());
+                }
             }
+    }
+
+    private boolean checkBind(FileStorage fileStorage) {
+        return fileStorageRepository.findAllByUuid(fileStorage.getUuid()).size() > 1;
     }
 
     private Map<Long, String> findFilesByEstimationAndFolder(Estimation estimation, Folder folder) {
